1. Добавил боссбар Bossbar.cs 2. Добавил арену для боссов BattleArena.cs 3. Доделал костер (полностью) Bonfire.cs 4. Сделал смену суток (и солнце) Sun.cs 5. Сделал зажигалку Firestarter.cs

+++ Assets/Scripts/Items/Data/BurnableItemData.cs
@@ -5,5 +5,5 @@ public class BurnableItemData : ItemData
 {
     [Header("BurnableItemData")]
     [Header("Длительность горения (в игровых минутах)")]
-    public float burningDuration = 30;
+    public int burningDuration = 30;
 }
